// src/app/generated_services/model/showLessonScheduleDTO.ts
/**
 * CarlsonGracie Backend API
 * API for managing students, belts, graduation requirements, notices, frequencies and user authentication with Keycloak
 *
 * Contact: marcosturossi@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { DayOfWeek } from './dayOfWeek';


export interface ShowLessonScheduleDTO { 
    id?: string;
    createdAt?: string;
    title: string;
    description?: string;
    dayOfWeek: DayOfWeek;
    startTime: string;
    duration: string;
    isActive?: boolean;
}
export namespace ShowLessonScheduleDTO {
}